+++ CHANGELOG.md
@@ -88,6 +88,7 @@
 - Reduced the show-more functionality on activity posts !2401 @sefn
 - Fix the list of nearby baskets based on the user's home location !1908 @alex.simm
 - Show date inn preview of voting module !1408 @Friedrich-B
+- CreateEditStoreCept, should run smoother !2433 @sefn
 
 ## Refactoring
 - Currently running events are shown within the event overview until their end !2162 @merhoc